* 'antdev' of https://gitlab.com/algaeh/algaeh-monorepo:
  attendence issuse solved
  Added confirm in delete notifications
  Added confirm in delete notifications

+++ client/src/Components/common/Notifications/notificationList.jsx
@@ -7,6 +7,7 @@ import {
   Avatar,
   Spin,
   AlgaehMessagePop,
+  Modal,
 } from "algaeh-react-components";
 import newAlgaehApi from "../../../hooks/newAlgaehApi";
 import { Link } from "react-router-dom";
@@ -90,27 +91,33 @@ export default function NotificationList({
       console.error(e);
     }
   }
-  async function callDeleteAllNotifications() {
-    try {
-      setLoading(true);
-      await newAlgaehApi({
-        uri: "/deleteAllNotification",
-        module: "documentManagement",
-        method: "DELETE",
-        data: {
-          user_id: userToken.employee_id,
-        },
-      });
-      callNotifications({ require_total_count: true });
-      AlgaehMessagePop({
-        type: "success",
-        display: "Successfully deleted all notifications",
-      });
-    } catch (e) {
-      AlgaehMessagePop({ type: "error", display: e.message });
-    } finally {
-      setLoading(false);
-    }
+  function callDeleteAllNotifications() {
+    Modal.confirm({
+      title: "Do you want to delete all notifications ?",
+      content: "Deleted notifications can not be restore again.",
+      onOk: async () => {
+        try {
+          setLoading(true);
+          await newAlgaehApi({
+            uri: "/deleteAllNotification",
+            module: "documentManagement",
+            method: "DELETE",
+            data: {
+              user_id: userToken.employee_id,
+            },
+          });
+          await callNotifications({ require_total_count: true });
+          AlgaehMessagePop({
+            type: "success",
+            display: "Successfully deleted all notifications",
+          });
+        } catch (e) {
+          AlgaehMessagePop({ type: "error", display: e.message });
+        } finally {
+          setLoading(false);
+        }
+      },
+    });
   }
   function showTotal() {
     return `Total ${totalRecords} items`;
